[CrowdSec]: Refactor to use provided templates
