+++ external-import/crowdsec/.flake8
@@ -2,3 +2,4 @@
 ignore = E203, E266, E501, W503, F403, F401
 max-line-length = 120
 select = B,C,E,F,W,T4,B9
+exclude = venv
