+++ external-import/crowdsec/entrypoint.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 # Correct working directory
-cd /opt/opencti-crowdsec-import || exit
+cd /opt/opencti-crowdsec || exit
 
 # Start the connector
 python3 main.py
