+++ external-import/crowdsec/src/requirements.txt
@@ -1,2 +1,4 @@
-pycti==6.7.4
-pycountry
+pycti==6.7.3
+pycountry==24.6.1
+stix2==3.0.1
+PyYAML==6.0.2
