+++ external-import/crowdsec/tests/resources/stix_observable.json
